+++ app/models/maths_and_physics/admin_tasks_presenter.rb
@@ -24,6 +24,13 @@ module MathsAndPhysics
       ]
     end
 
+    def identity_confirmation
+      [
+        ["Current school", eligibility.current_school.name],
+        ["Contact number", eligibility.current_school.phone_number]
+      ]
+    end
+
     private
 
     def eligibility