+++ spec/features/admin_checks_maths_and_physics_claim_spec.rb
@@ -88,4 +88,37 @@ RSpec.feature "Admin checking a Maths & Physics claim" do
     expect(page).to have_content("You must select ‘Yes’ or ‘No’")
     expect(claim.tasks.find_by(name: "qualifications")).to be_nil
   end
+
+  scenario "service operator checks and approves a Maths & Physics claim where the claimant did not complete GOV.UK Verify" do
+    claim = create(:claim, :unverified, policy: MathsAndPhysics)
+
+    visit admin_claims_path
+    find("a[href='#{admin_claim_tasks_path(claim)}']").click
+
+    expect(page).to have_content("1. Qualifications")
+    expect(page).to have_content("2. Employment")
+    expect(page).to have_content("3. Identity confirmation")
+    expect(page).to have_content("4. Decision")
+
+    click_on I18n.t("maths_and_physics.admin.tasks.identity_confirmation.summary")
+
+    expect(page).to have_content(I18n.t("maths_and_physics.admin.tasks.identity_confirmation.question", name: claim.full_name))
+    expect(page).to have_content(claim.eligibility.current_school.name)
+    expect(page).to have_content(claim.eligibility.current_school.phone_number)
+
+    choose "Yes"
+    click_on "Save and continue"
+
+    expect(claim.tasks.find_by!(name: "identity_confirmation").passed?).to eq(true)
+
+    expect(page).to have_content("Claim decision")
+
+    choose "Approve"
+    fill_in "Decision notes", with: "All checks passed!"
+    click_on "Confirm decision"
+
+    expect(page).to have_content("Claim has been approved successfully")
+    expect(claim.decision).to be_approved
+    expect(claim.decision.created_by).to eq(user)
+  end
 end